extract menu bar utils into pom
